+++ Gruntfile.js
@@ -189,6 +189,7 @@ module.exports = function (grunt) {
     ]);
 
     grunt.registerTask('default', [
+        'production',
     ]);
 
     grunt.registerTask('publish', [